requirements.txt
torch
torchvision
gradio
pillow
git+https://github.com/huggingface/diffusers
transformers
accelerate
sentencepiece
